D_draw.key {lattice}	R Documentation

Produce a Legend or Key

Description

Produces (and possibly draws) a Grid frame grob which is a legend (aka key) that can be placed in other Grid plots.

Usage

draw.key(key, draw=FALSE, vp=NULL, ...)

Arguments

key
A list determining the key. See documentation for xyplot, in particular the section describing the key argument, for details.

draw
logical, whether the grob is to be drawn.

vp
viewport

...
ignored

Value

A Grid frame object (that inherits from ‘grob’).

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

xyplot

[Package lattice version 0.20-33 Index]