llines {lattice}	R Documentation

Lattice Replacements of base graphics functions

Description

These functions are intended to replace some commonly used base R graphics functions in panel functions.

Usage

larrows(x0, y0, x1, y1, length = 0.25, unit = "inches", ...)
llines(x, y, ...)
lplot.xy(xy, ...)
lpoints(x, ...)
lsegments(x0, y0, x1, y1, ...)
lsegments(x1, y1, x2, y2, ...)
ltext(x, ...)
panel.arrows(...)
panel.lines(...)
panel.points(...)
panel.segments(...)
panel.text(...)

Arguments

x, y, x0, y0, x1, y1, x2, y2, xy 	 locations. x2, y2 present for S compatibility

length, unit 	 determines extent of arrow head. length specifies the length in terms of unit, which can be any valid grid unit as long as it doesn't need a data argument. unit defaults to inches, which is the only option in the base version of the function, arrows.

... 	 other arguments, including ones to control graphical parameters like color, line type, etc. For devices that support alpha-transparency, a numeric argument alpha between 0 and 1 can be supplied. Be careful with this, since for devices that do not support alpha-transparency, nothing will be drawn at all if this is set to anything other than 0.

Details

These functions are meant to be grid replacements of the corresponding base R graphics functions, to allow existing Trellis code to be used with minimal modification. The functions panel.* are essentally identical to the l* versions, are recommended for use in new code (as opposed to ported code) as they have more readable names.

See the documentation of the base functions for usage. Not all arguments are always supported. All these correspond to the default methods only. For ltext, only values 0, .5 and 1 for adj have any effect.

Note

There's a new type="H" option wherever appropriate, which is similar to type="h", but with horizontal lines.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

points, lines, text, segments, arrows, Lattice

[Package lattice version 0.11-6 Index]