panel.bwplot {lattice}	R Documentation

Default Panel Function for bwplot

Description

This is the default panel function for bwplot.

Usage

panel.bwplot(x, y,
             box.ratio = 1, horizontal,
             pch, col, cex, fill = "transparent",
             varwidth = FALSE, coef = 1.5, do.out = TRUE, ...)

Arguments

x, y 	 numeric vector or factor. Boxplots drawn for each unique value of y (x) if horizontal is TRUE (FALSE)

box.ratio 	 ratio of box height to inter box space

horizontal 	 logical. If FALSE, the plot is `transposed' in the sense that the behaviours of x and y are switched. x is now the `factor'. Interpretation of other arguments change accordingly. See documentation of bwplot for a fuller explanation.

pch, col, cex 	 graphical parameters controlling the dot

fill 	 color to fill the boxplot

varwidth 	 logical. If TRUE, widths of boxplots are proportional to the number of points used in creating it.

coef, do.out 	 passed to boxplot.stats

... 	 more graphical parameters controlling the display can be supplied

Details

Creates Box and Whisker plot of x for every level of y. The actual boxplot statistics are calculated using boxplot.stats. Note that most arguments controlling the display can be supplied to the high-level bwplot call directly.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

bwplot, boxplot.stats

[Package lattice version 0.11-6 Index]