F_1_panel.histogram {lattice}	R Documentation

Default Panel Function for histogram

Description

This is the default panel function for histogram.

Usage

panel.histogram(x,
                breaks,
                equal.widths = TRUE,
                type = "density",
                nint = round(log2(length(x)) + 1), 
                alpha, col, border, lty, lwd,
                ...,
                identifier = "histogram")

Arguments

x
The data points for which the histogram is to be drawn

breaks
The breakpoints for the histogram

equal.widths
logical used when breaks==NULL

type
Type of histogram, possible values being "percent", "density" and "count"

nint
Number of bins for the histogram

alpha, col, border, lty, lwd
graphical parameters for bars; defaults are obtained from the plot.polygon settings.

...
other arguments, passed to hist when deemed appropriate

identifier
A character string that is prepended to the names of grobs that are created by this panel function.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

histogram

[Package lattice version 0.20-33 Index]