F_2_panel.loess {lattice}	R Documentation

Panel Function to Add a LOESS Smooth

Description

A predefined panel function that can be used to add a LOESS smooth based on the provided data.

Usage



panel.loess(x, y, span = 2/3, degree = 1,
            family = c("symmetric", "gaussian"),
            evaluation = 50,
            lwd, lty, col, col.line, type,
            horizontal = FALSE,
            ..., identifier = "loess")

Arguments

x, y
Variables defining the data to be used.

lwd, lty, col, col.line
Graphical parameters for the added line. col.line overrides col.

type
Ignored. The argument is present only to make sure that an explicitly specified type argument (perhaps meant for another function) does not affect the display.

span, degree, family, evaluation
Arguments to loess.smooth, for which panel.loess is essentially a wrapper.

horizontal
A logical flag controlling which variable is to be treated as the predictor (by default x) and which as the response (by default y). If TRUE, the plot is ‘transposed’ in the sense that y becomes the predictor and x the response. (The name ‘horizontal’ may seem an odd choice for this argument, and originates from similar usage in bwplot).

...
Extra arguments, passed on to panel.lines.

identifier
A character string that is prepended to the names of grobs that are created by this panel function.

Value

The object returned by loess.smooth.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

Lattice, loess.smooth, prepanel.loess

[Package lattice version 0.20-35 Index]