panel.stripplot {lattice}	R Documentation

Default Panel Function for stripplot

Description

This is the default panel function for stripplot. Also see panel.superpose

Usage

panel.stripplot(x, y, jitter.data = FALSE, factor = 0.5,
                horizontal = TRUE, groups = NULL,
                ...)

Arguments

x,y 	 coordinates of points to be plotted

jitter.data 	 whether points should be jittered to avoid overplotting.

factor 	 factor by which to jitter

horizontal 	 logical. If FALSE, the plot is `transposed' in the sense that the behaviours of x and y are switched. x is now the `factor'. Interpretation of other arguments change accordingly. See documentation of bwplot for a fuller explanation.

groups 	 optional grouping variable

... 	 graphical parameters etc can be supplied, passed to panel.xyplot or panel.superpose, depending on whether groups is present

Details

Creates stripplot (one dimensional scatterplot) of x for each level of y

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

stripplot

[Package lattice version 0.11-6 Index]