panel.xyplot {lattice}	R Documentation

Default Panel Function for xyplot

Description

This is the default panel function for xyplot. Also see panel.superpose. The default panel functions for splom and qqmath are essentially the same function.

Usage

panel.xyplot(x, y, type="p",
             pch = plot.symbol$pch,
             col,
             col.line = plot.line$col,
             col.symbol = plot.symbol$col,
             font = plot.symbol$font,
             fontfamily = plot.symbol$fontfamily,
             fontface = plot.symbol$fontface,
             lty = plot.line$lty,
             cex = plot.symbol$cex,
             lwd = plot.line$lwd,
             horizontal = FALSE, ...)
panel.splom(...)
panel.qqmath(...)
panel.qq(...)

Arguments

x,y 	 variables to be plotted in the scatterplot

type 	 character vector consisting of one or more of the following: "p", "l", "h", "b", "o", "s", "S", "r", "g", "smooth". If type has more than one element, an attempt is made to combine the effect of each of the components.
The behaviour if any of the first six are included in type is similar to the effect of type in plot (type 'b' is actually the same as 'o'). "r" adds a regression line (same as panel.lmline, except for default graphical parameters), and "smooth" adds a lowess fit (same as panel.loess). "g" adds a reference grid using panel.grid in the background.
See example(xyplot) and demo(lattice) for examples.

col, col.line, col.symbol 	 default colours are obtained from plot.symbol and plot.line using trellis.par.get.

font, fontface, fontfamily 	 font used when pch is a character

pch, lty, cex, lwd 	 other graphical parameters.

... 	 extra arguments, if any, for panel.xyplot. In most cases panel.xyplot ignores these. For types "r" and "smooth", these are passed on to panel.lmline and panel.loess respectively.

horizontal 	 logical. Controls orientation when type is one of "h", "s" or "S"

Details

Creates scatterplot of x and y, with various modifications possible via the type argument. panel.qq draws a 45 degree line before calling panel.xyplot.

Note that most of the arguments controlling the display can be supplied directly to the high-level xyplot call.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

panel.superpose, xyplot, splom,qqmath

[Package lattice version 0.11-6 Index]