qq {lattice}	R Documentation

Quantile-Quantile Plots of Two Samples

Description

Quantile-Quantile plots for comparing two Distributions

Usage

qq(formula, f.value = NULL, ...)

Arguments

formula 	 formula of the form y ~ x | g1 * g2 * ..., where x must be a numeric, and y can be a factor, shingle, character or numeric vector, with the restriction that there must be exactly two levels of y, which divide the values of x into two groups. Quantiles for these groups will be plotted along the two axes.

f.value 	 optional function of a single integer (representing the larger of the two sample sizes), returning a vector of probabilities corresponding to which quantiles should be plotted. The typical value for this argument is ppoints, which is also the S-PLUS default. If specified, the probabilities generated by this function is used for the plotted quantiles, using the quantile function.
f.value defaults to NULL, which is equivalent to using function(n) ppoints(n, a = 1). This has the effect of including the minimum and maximum data values in the computed quantiles. This is similar to what happens for qqplot but different from the default qq behaviour in S-PLUS.

... 	 Other arguments

Details

qq produces a Q-Q plot of two samples. The default behaviour of qq is different from the corresponding S-PLUS function. See the entry for f.value for specifics.

This and all other high level Trellis functions have several arguments in common. These are extensively documented only in the help page for xyplot, which should be consulted to learn more detailed usage.

Value

An object of class ``trellis''. The `update' method can be used to update components of the object and the `print' method (usually called by default) will plot it on an appropriate plotting device.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

xyplot, panel.qq, qqmath, Lattice

Examples

qq(voice.part ~ height, aspect = 1, data = singer,
   subset = (voice.part == "Bass 2" | voice.part == "Tenor 1"))

[Package lattice version 0.11-6 Index]