strip.default {lattice}	R Documentation

In-built Trellis Strip Function

Description

strip.default is the function that draws the strips by default in conditioning Trellis plots. Users can write their own strip functions, but most commonly this involves calling strip.default with a slightly different arguments. strip.custom provides a convenient way to obtain new strip functions that differ from strip.default only in the default values of certain arguments.

Usage

strip.default(which.given,
              which.panel,
              var.name,
              factor.levels,
              shingle.intervals,
              strip.names = c(FALSE, TRUE),
              style = 1,
              bg = trellis.par.get("strip.background")$col[which.given],
              fg = trellis.par.get("strip.shingle")$col[which.given],
              par.strip.text = trellis.par.get("add.text"))
strip.custom(...)

Arguments

which.given 	 integer index specifying which of the conditioning variables this strip corresponds to.

which.panel 	 vector of integers as long as the number of conditioning variables. The contents are indices specifing the current levels of each of the conditioning variables (thus, this would be unique for each distinct panel).

var.name 	 vector of character strings or ecpressions as long as the number of conditioning variables. The contents are the names of the conditioning variables, to be used if the appropriate entry in strip.names (see below) is true.

factor.levels 	 if the current strip corresponds to a factor, this should be a character or expression vector giving the levels of the factor. Otherwise, it should be NULL

shingle.intervals 	 if the current strip corresponds to a shingle, this should be a 2-column matrix giving the levels of the shingle. (of the form that would be produced by printing levels(shingle)). Otherwise, it should be NULL

strip.names 	 a logical vector of length 2, indicating whether or not the name of the conditioning variable that corresponds to the strip being drawn is to be written on the strip. The two components give the values for factors and shingles respectively.
This argument is ignored for a factor when style is not one of 1 and 3.

style 	 integer, with values 1,2,3,4,5 and 6 currently supported. Applicable only when x is a factor. Determines how the current level of x is indicated on the strip.
The best way to find out what effect the value of style has is to try them out. Here is a short description: for a style value of 1, the strip is colored in the background color with the strip text (as determined by other arguments) centred on it. A value of 3 is the same, except that a part of the strip is colored in the foreground color, indicating the current level of the factor. For styles 2 and 4, the part corresponding to the current level remains colored in the foreground color, however, for style = 2, the remaining part is not colored at all, whereas for 4, it is colored with the background color. For both these, the names of all the levels of the factor are placed on the strip from left to right. Styles 5 and 6 produce the same effect (they are subtly different in S, this implementation corresponds to 5), they are similar to style 1, except that the strip text is not centred, it is instead postioned according to the current level.
Note that unlike S-PLUS, the default value of style is 1.

par.strip.text 	 list with parameters controlling the text on each strip, with components col, cex, font, etc.

bg 	 strip background color.

fg 	 strip foreground color.

... 	 arguments to be passed on to strip.default, overriding whatever value it would have normally assumed

Details

default strip function for trellis functions. Useful mostly because of the style argument — non-default styles are often more informative, especially when the names of the levels of the factor x are small. Typical use is as strip = function(...) strip.default(style=2,...).

Value

strip.default is useful for its side-effect, which is to draw a strip appropriate for conditioning Trellis plots. strip.default returns a function similar to strip.default

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

xyplot, Lattice

Examples

xyplot(Petal.Length ~ Petal.Width | Species, iris,
       strip = strip.custom(style = 4))

[Package lattice version 0.11-6 Index]