G_utilities.3d {lattice}	R Documentation

Utility functions for 3-D plots

Description

These are (related to) the default panel functions for cloud and wireframe.

Usage

ltransform3dMatrix(screen, R.mat)
ltransform3dto3d(x, R.mat, dist)

Arguments

x
x can be a numeric matrix with 3 rows for ltransform3dto3d

screen
list, as described in panel.cloud

R.mat
4x4 transformation matrix in homogeneous coordinates

dist
controls transformation to account for perspective viewing

Details

ltransform3dMatrix and ltransform3dto3d are utility functions to help in computation of projections. These functions are used inside the panel functions for cloud and wireframe. They may be useful in user-defined panel functions as well.

The first function takes a list of the form of the screen argument in cloud and wireframe and a R.mat, a 4x4 transformation matrix in homogeneous coordinates, to return a new 4x4 transformation matrix that is the result of applying R.mat followed by the rotations in screen. The second function applies a 4x4 transformation matrix in homogeneous coordinates to a 3xn matrix representing points in 3-D space, and optionally does some perspective computations. (There has been no testing with non-trivial transformation matrices, and my knowledge of the homogeneous coordinate system is very limited, so there may be bugs here.)

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

cloud, panel.cloud

[Package lattice version 0.20-33 Index]