banking {lattice}	R Documentation

Banking

Description

Calculates `banking' slope

Usage

banking(dx, dy)

Arguments

dx, dy 	 vector of consecutive x, y differences.

Details

banking is a crude banking function used when aspect="xy" in high level Trellis functions. Its usually not very menaingful except for xyplot. It just orders the absolute slopes and returns a value which when adjusted by the panel scale limits will make the median of the above absolute slopes a 45 degree line.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

Lattice, xyplot

[Package lattice version 0.11-6 Index]