banking {lattice}	R Documentation

Banking

Description

Calculates banking slope

Usage

banking(dx, dy)

Arguments

dx, dy 	 vector of consecutive x, y differences.

Details

banking is the banking function used when aspect = "xy" in high level Trellis functions. It is usually not very meaningful except with xyplot. It considers the absolute slopes (based on dx and dy) and returns a value which when adjusted by the panel scale limits will make the median of the above absolute slopes correspond to a 45 degree line.

This function was inspired by the discussion of banking in the documentation for Trellis Graphics available at Bell Labs' website (see Lattice), but is most likely identical to an algorithm described by Cleveland et al (see below). It is not clear (to the author) whether this is the algorithm used in S-PLUS. Alternative banking rules, implemented as a similar function, can be used as a drop-in replacement by suitably modifying lattice.options("banking").

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

References

Cleveland, William S., McGill, Marylyn E. and McGill, Robert (1988), "The Shape Parameter of a Two-variable Graph", Journal of the American Statistical Association, 83, 289-300

See Also

Lattice, xyplot

Examples

xyplot(sunspot.year ~ time(sunspot.year) | equal.count(time(sunspot.year)), 
       xlab = "", type = "l", aspect = "xy", strip = FALSE,
       scales = list(x = list(alternating = 2, relation = "sliced")),
       as.table = TRUE, main = "Yearly Sunspots")

[Package lattice version 0.17-25 Index]