draw.colorkey {lattice}	R Documentation

Produce a Colorkey for levelplot

Description

Produces (and possibly draws) a Grid frame grob which is a colorkey that can be placed in other Grid plots. Used in levelplot

Usage

draw.colorkey(key, draw=FALSE, vp=NULL)

Arguments

key 	 A list determining the key. See documentation for levelplot, in particular the section describing the colorkey argument, for details.

draw 	 logical, whether the grob is to be drawn.

vp 	 viewport

Value

A Grid frame object (that inherits from "grob")

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

xyplot

[Package lattice version 0.17-25 Index]