lattice.options {lattice}	R Documentation

Low-level Options Controlling Behaviour of Lattice

Description

Functions to handle settings used by lattice. Their main purpose is to make code maintainance easier, and users normally should not need to use these functions. However, fine control at this level maybe useful in certain cases.

Usage

lattice.options(...)
lattice.getOption(name)

Arguments

name 	 character giving the name of a setting

... 	 new options can be defined, or existing ones modified, using one or more arguments of the form name = value or by passing a list of such tagged values. Existing values can be retrieved by supplying the names (as character strings) of the components as unnamed arguments.

Details

These functions are modeled on options and getOption, and behave similarly for the most part. The components currently used are not documented here, but are fairly self-explanatory.

Value

lattice.getOption returns the value of a single component, whereas lattice.options always returns a list with one or more named components. When changing the values of components, the old values of the modified components are returned by lattice.options. If called without any arguments, the full list is returned.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

options, trellis.device, trellis.par.get, Lattice

Examples

str(lattice.options())
lattice.getOption("save.object")

[Package lattice version 0.17-25 Index]