lset {lattice}	R Documentation

Interface to modify Trellis Settings - Deprecated

Description

A (hopefully) simpler alternative to trellis.par.get/set. This is deprecated, and the same functionality is now available with trellis.par.set

Usage

lset(theme = col.whitebg())

Arguments

theme 	 a list decribing how to change the settings of the current active device. Valid components are those in the list returned by trellis.par.get(). Each component must itself be a list, with one or more of the appropriate components (need not have all components). Changes are made to the settings for the currently active device only.

Details

lset is a wrapper to trellis.par.set that allows setting more than one parameter at a time, and only the components that are to be changed need be specified.

This function is expected to be used in conjunction with functions that return `themes', i.e., lists which when supplied to lset change the overall look and feel of the display.

lset is used to modify the settings on a one-time basis. It is also possible to make persistent changes in the default theme used via a user-settable option, see trellis.device for details.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

trellis.device, show.settings, Lattice

Examples

show.settings(col.whitebg())
# The following reproduces S-PLUS Unix black and white pch:
#lset(list(superpose.symbol=list(pch=c("o","+",">","s","w","#","{"))))

[Package lattice version 0.11-6 Index]