lset {lattice}	R Documentation

Interface to modify Trellis Settings - Deprecated

Description

A (hopefully) simpler alternative to trellis.par.get/set. This is deprecated, and the same functionality is now available with trellis.par.set

Usage

lset(theme = col.whitebg())

Arguments

theme 	 a list decribing how to change the settings of the current active device. Valid components are those in the list returned by trellis.par.get(). Each component must itself be a list, with one or more of the appropriate components (need not have all components). Changes are made to the settings for the currently active device only.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

[Package lattice version 0.17-25 Index]