oneway {lattice}	R Documentation

Fit One-way Model

Description

Fits a One-way model to univariate data grouped by a factor, the result often being displayed using rfs

Usage

oneway(formula, data, location=mean, spread=function(x) sqrt(var(x)))

Arguments

formula 	 formula of the form y ~ x where y is the numeric response and x is the grouping factor

data 	 data frame in which the model is to be evaluated

location 	 function or numeric giving the location statistic to be used for centering the observations, e.g. median, 0 (to avoid centering).

spread 	 function or numeric giving the spread statistic to be used for scaling the observations, e.g. sd, 1 (to avoid scaling).

Value

A list with components

location 	 vector of locations for each group.

spread 	 vector of spreads for each group.

fitted.values 	 vector of locations for each observation.

residuals 	 residuals (y - fitted.values).

scaled.residuals 	 residuals scaled by spread for their group

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

rfs, Lattice

[Package lattice version 0.17-25 Index]