panel.axis {lattice}	R Documentation

Panel Function for Drawing Axis Ticks and Labels

Description

panel.axis is the function used by lattice to draw axes. It is typically not used by users, except those wishing to create advanced annotation. Keep in mind issues of clipping when trying to use it as part of the panel function. current.panel.limits can be used to retrieve a panel's x and y limits.

Usage

panel.axis(side = c("bottom", "left", "top", "right"),
           at,
           labels = TRUE,
           draw.labels = TRUE,
           check.overlap = FALSE,
           outside = FALSE,
           ticks = TRUE,
           half = !outside,
           which.half,
           tck = as.numeric(ticks),
           rot = if (is.logical(labels)) 0 else c(90, 0),
           text.col, text.alpha, text.cex, text.font,
           text.fontfamily, text.fontface,
           line.col, line.lty, line.lwd, line.alpha)

current.panel.limits(unit = "native")

Arguments

side 	 character string indicating which side axes are to be drawn on. Partial specification is allowed.

at 	 location of labels

labels 	 the labels to go along with at. The labels can be a character vector or a vector of expressions. Alternatively, this can be a logical. If TRUE, the labels are derived from at, otherwise, labels are empty.

draw.labels 	 logical indicating whether labels are to be drawn

check.overlap 	 logical, whether to check for overlapping of labels. This also has the effect of removing at values that are ‘too close’ to the limits.

outside 	 logical, whether to the labels draw outside the panel or inside. Note that outside=TRUE will only have a visible effect if clipping is disabled for the viewport (panel).

ticks 	 logical, whether to draw the tickmarks

half 	 logical, whether only half of scales will be drawn for each side

which.half 	 character string, one of "lower" and "upper". Indicates which half is to be used for tick locations if half=TRUE. Defaults to whatever is suitable for splom

tck 	 numeric scalar, multiplier for tick length. Can be negative.

rot 	 rotation angles for labels in degrees. Can be a vector of length 2 for x- and y-axes respectively

text.col, text.alpha, text.cex, text.font, text.fontfamily, text.fontface, line.col, line.lty, line.lwd, line.alpha 	 graphical parameters

unit 	 which grid unit the values should be in

Details

panel.axis can draw axis tick marks inside or outside a panel (more precisely, a grid viewport). It honours the (native) axis scales. Used in panel.pairs for splom, as well as for all the usual axis drawing by the print method for "trellis" objects. It can also be used to enhance plots ‘after the fact’ by adding axes.

Value

current.panel.limits returns a list with components xlim and ylim, which are both numeric vectors of length 2, giving the scales of the current panel (viewport). The values correspond to the unit system specified by unit, by default "native".

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

Lattice, xyplot, trellis.focus, unit

[Package lattice version 0.17-25 Index]