panel.densityplot {lattice}	R Documentation

Default Panel Function for densityplot

Description

This is the default panel function for densityplot.

Usage

panel.densityplot(x, darg, plot.points = TRUE, ref = FALSE,
                  col, col.line, ...)

Arguments

x 	 data points for which density is to be estimated

darg 	 list of arguments to be passed to the density function. Typically, this should be a list with zero or more of the following components : bw, adjust, kernel, window, width, give.Rkern, n, from, to, cut, na.rm (see density for details)

plot.points 	 logical, specifying whether or not the data points should be plotted along with the estimated density.

ref 	 logical, whether to draw x-axis

col 	 color for points and line

col.line 	 color for line, overrides col

... 	 extra grahical parameters

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

densityplot

[Package lattice version 0.11-6 Index]