panel.densityplot {lattice}	R Documentation

Default Panel Function for densityplot

Description

This is the default panel function for densityplot.

Usage

panel.densityplot(x, darg, plot.points = "jitter", ref = FALSE,
                  groups = NULL, weights = NULL,
                  jitter.amount, type, ...)

Arguments

x 	 data points for which density is to be estimated

darg 	 list of arguments to be passed to the density function. Typically, this should be a list with zero or more of the following components : bw, adjust, kernel, window, width, give.Rkern, n, from, to, cut, na.rm (see density for details)

plot.points 	 logical specifying whether or not the data points should be plotted along with the estimated density. Alternatively, a character string specifying how the points should be plotted. Meaningful values are "rug", in which case panel.rug is used to plot a ‘rug’, and "jitter", in which case the points are jittered vertically to better distinguish overlapping points.

ref 	 logical, whether to draw x-axis

groups 	 an optional grouping variable. If present, panel.superpose will be used instead to display each subgroup

weights 	 numeric vector of weights for the density calculations. If this is specified, the ... part must also include a subscripts argument that matches the weights to x.

jitter.amount 	 when plot.points="jitter", the value to use as the amount argument to jitter.

type 	 type argument used to plot points, if requested. This is not expected to be useful, it is available mostly to protect a type argument, if specified, from affecting the density curve.

... 	 extra graphical parameters. Note that additional arguments to panel.rug cannot be passed on through panel.densityplot.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

densityplot, jitter

[Package lattice version 0.17-25 Index]