panel.dotplot {lattice}	R Documentation

Default Panel Function for dotplot

Description

Default panel function for dotplot.

Usage

panel.dotplot(x, y, horizontal = TRUE,
              pch, col, lty, lwd, 
              col.line, levels.fos,
              groups = NULL,
              ...)

Arguments

x,y 	 variables to be plotted in the panel. Typically y is the ‘factor’

horizontal 	 logical. If FALSE, the plot is ‘transposed’ in the sense that the behaviours of x and y are switched. x is now the ‘factor’. Interpretation of other arguments change accordingly. See documentation of bwplot for a fuller explanation.

pch, col, lty, lwd, col.line 	 graphical parameters

levels.fos 	 locations where reference lines will be drawn

groups 	 grouping variable (affects graphical parameters)

... 	 extra parameters, passed to panel.xyplot which is responsible for drawing the foreground points (panel.dotplot only draws the background reference lines).

Details

Creates (possibly grouped) Dotplot of x against y or vice versa

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

dotplot

[Package lattice version 0.17-25 Index]