panel.levelplot {lattice}	R Documentation

Default Panel Function for levelplot

Description

This is the default panel function for levelplot.

Usage

panel.levelplot(x, y, z, 
                subscripts,
                at = pretty(z),
                shrink,
                labels,
                label.style = c("mixed", "flat", "align"),
                contour = FALSE,
                region = TRUE,
                col = add.line$col,
                lty = add.line$lty,
                lwd = add.line$lwd,
                ...,
                col.regions = regions$col,
                alpha.regions = regions$alpha)
panel.contourplot(...)

Arguments

x, y, z 	 Variables defining the plot.

subscripts 	 Integer vector indicating what subset of x, y and z to draw.

at 	 Numeric vector giving breakpoints along the range of z. See levelplot for details.

shrink 	 Either a numeric vector of length 2 (meant to work as both x and y components), or a list with components x and y which are numeric vectors of length 2. This allows the rectangles to be scaled proportional to the z-value. The specification can be made separately for widths (x) and heights (y). The elements of the length 2 numeric vector gives the minimum and maximum proportion of shrinkage (corresponding to min and max of z).

labels 	 Either a logical scalar indicating whether the labels are to be drawn, or a character or expression vector giving the labels associated with the at values. Alternatively, labels can be a list with the following components:

labels:

a character or expression vector giving the labels. This can be omitted, in which case the defaults will be used.

col, cex, alpha:

graphical parameters for label texts

fontfamily, fontface, font:

font used for the labels

label.style 	 Controls how label positions and rotation are determined. A value of "flat" causes the label to be positioned where the contour is flattest, and the label is not rotated. A value of "align" causes the label to be drawn as far from the boundaries as possible, and the label is rotated to align with the contour at that point. The default is to mix these approaches, preferring the flattest location unless it is too close to the boundaries.

contour 	 A logical flag, specifying whether contour lines should be drawn.

region 	 A logical flag, specifying whether inter-contour regions should be filled with the appropriate color.

col, lty, lwd 	 graphical parameters for contour lines

... 	 Extra parameters.

col.regions 	 A vector of colors, or a function to produce a vecor of colors, to be used if region=TRUE. Each interval defined by at is assigned a color, so the number of colors actually used is one less than the length of at. See level.colors for details on how the color assignment is done.

alpha.regions 	 numeric scalar controlling transparency of facets

Details

The same panel function is used for both levelplot and contourplot (which differ only in default values of some arguments). panel.contourplot is a simple wrapper to panel.levelplot.

When contour=TRUE, the contourLines function is used to calculate the contour lines.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

levelplot, level.colors, contourLines

[Package lattice version 0.17-25 Index]