panel.pairs {lattice}	R Documentation

Default Superpanel Function for splom

Description

This is the default superpanel function for splom.

Usage

panel.pairs(z,
            panel = lattice.getOption("panel.splom"),
            lower.panel = panel,
            upper.panel = panel,
            diag.panel = "diag.panel.splom",
            as.matrix = FALSE,
            groups = NULL,
            panel.subscripts,
            subscripts,
            pscales = 5,
            prepanel.limits = function(x) if (is.factor(x)) levels(x) else
            extend.limits(range(as.numeric(x), finite = TRUE)),

            varname.col, varname.cex, varname.font,
            varname.fontfamily, varname.fontface,
            axis.text.col, axis.text.cex, axis.text.font,
            axis.text.fontfamily, axis.text.fontface,
            axis.line.col, axis.line.lty, axis.line.lwd,
            axis.line.alpha, axis.line.tck,
            ...)
diag.panel.splom(x = NULL,
                 varname = NULL, limits, at = NULL, lab = NULL,
                 draw = TRUE,
                 varname.col, varname.cex,
                 varname.lineheight, varname.font,
                 varname.fontfamily, varname.fontface,
                 axis.text.col, axis.text.alpha,
                 axis.text.cex, axis.text.font, 
                 axis.text.fontfamily, axis.text.fontface,
                 axis.line.col, axis.line.alpha,
                 axis.line.lty, axis.line.lwd,
                 axis.line.tck,
                 ...)

Arguments

z 	 The data frame used for the plot.

panel, lower.panel, upper.panel 	 The panel function used to display each pair of variables. If specified, lower.panel and upper.panel are used for panels below and above the diagonal respectively.

diag.panel 	 The panel function used for the diagonals. See arguments to diag.panel.splom to know what arguments this function is passed when called.

as.matrix 	 logical. If TRUE, the layout of the panels will have origin on the top left instead of bottom left (similar to pairs). This is in essence the same functionality as provided by as.table for the panel layout

groups 	 Grouping variable, if any

panel.subscripts 	 logical specifying whether the panel function accepts an argument named subscripts.

subscripts 	 The indices of the rows of z that are to be displayed in this (super)panel.

pscales 	 Controls axis labels, passed down from splom. If pscales is a single number, it indicates the approximate number of equally-spaced ticks that should appear on each axis. If pscales is a list, it should have one component for each column in z, each of which itself a list with the following valid components:
at: a numeric vector specifying tick locations
labels: character vector labels to go with at
limits: numeric 2-vector specifying axis limits (should be made more flexible at some point to handle factors)
These are specifications on a per-variable basis, and used on all four sides in the diagonal cells used for labelling. Factor variables are labelled with the factor names. Use pscales=0 to supress the axes entirely.

prepanel.limits 	 The ‘regular’ high level lattice plots such as xyplot use the prepanel function for deciding on axis limits from data. This function serves a similar function, and works on a per-variable basis, by calculating the limits, which can be overridden by the corresponding limits component in the pscales list.

x 	 data vector corresponding to that row / column (which will be the same for diagonal ‘panels’).

varname 	 (scalar) character string or expression that is to be written centred within the panel

limits 	 numeric of length 2, or, vector of characters, specifying the scale for that panel (used to calculate tick locations when missing)

at 	 locations of tick marks

lab 	 optional labels for tick marks

draw 	 logical, specifying whether to draw the tick marks and labels. If FALSE, only variable names are written

varname.col, varname.cex, varname.lineheight, varname.font, varname.fontfamily, varname.fontface 	 graphical parameters for the variable name in each diagonal panel

axis.text.col, axis.text.cex, axis.text.font, axis.text.fontfamily, axis.text.fontface, axis.line.col, axis.line.lty, axis.line.lwd, axis.text.alpha, axis.line.alpha 	 graphical parameters for axis tick marks and labels

axis.line.tck 	 length of tick marks in diagonal panels

... 	 extra arguments passed on to panel, lower.panel, upper.panel and diag.panel from panel.pairs. Currently ignored by diag.panel.splom.

Details

panel.pairs is the function that is actually passed in as the panel function in a trellis object produced by splom (taking the panel function as its argument).

Note that the axis labeling does not support date-time classes at present.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

splom

[Package lattice version 0.17-25 Index]