panel.parallel {lattice}	R Documentation

Default Panel Function for parallel

Description

This is the default panel function for parallel.

Usage

panel.parallel(x, y, z, subscripts,
               groups = NULL,
               col, lwd, lty, alpha,
               common.scale = FALSE,
               lower,
               upper,
               ...,
               horizontal.axis = TRUE)

Arguments

x, y 	 dummy variables, ignored.

z 	 The data frame used for the plot. Each column will be coerced to numeric before being plotted, and an error will be issued if this fails.

subscripts 	 The indices of the rows of z that are to be displyed in this panel.

groups 	 An optional grouping variable. If specified, different groups are distinguished by use of different graphical parameters (i.e., rows of z in the same group share parameters).

col, lwd, lty, alpha 	 graphical parameters (defaults to the settings for superpose.line). If groups is non-null, these parameters used one for each group. Otherwise, they are recycled and used to distinguish between rows of the data frame z.

common.scale 	 logical, whether a common scale should be used columns of z. Defaults to FALSE, in which case the horizontal range for each column is different (as determined by lower and upper).

lower, upper 	 numeric vectors replicated to be as long as the number of columns in z. Determines the lower and upper bounds to be used for scaling the corresponding columns of z after coercing them to numeric. Defaults to the minimum and maximum of each column. Alternatively, these could be functions (to be applied on each column) that return a scalar.

... 	 other arguments (ignored)

horizontal.axis 	 logical indicating whether the parallel axes should be laid out horizontally (TRUE) or vertically (FALSE).

Details

Produces parallel coordinate plots, which are easier to understand from an example than through a verbal description. See example for parallel

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

References

Inselberg, Alfred (2009) “Parallel Coordinates: Visual Multidimensional Geometry and Its Applications”, Springer. ISBN: 978-0-387-21507-5.

Inselberg, A. (1985) “The Plane with Parallel Coordinates”, The Visual Computer.

See Also

parallel

[Package lattice version 0.17-25 Index]