panel.qqmath {lattice}	R Documentation

Default Panel and Prepanel Function for qqmath

Description

This is the default panel function for qqmath.

Usage

panel.qqmath(x, f.value = NULL,
             distribution = qnorm,
             qtype = 7,
             groups = NULL, ...)

Arguments

x 	 vector (typically numeric, coerced if not) of data values to be used in the panel.

f.value, distribution 	 Defines how quantiles are calculated. See qqmath for details.

qtype 	 The type argument to be used in quantile

groups 	 An optional grouping variable. Within each panel, one Q-Q plot is produced for every level of this grouping variable, differentiated by different graphical parameters.

... 	 further arguments, often graphical parameters.

Details

Creates a Q-Q plot of the data and the theoretical distribution given by distribution. Note that most of the arguments controlling the display can be supplied directly to the high-level qqmath call.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

qqmath

[Package lattice version 0.17-25 Index]