panel.qqmathline {lattice}	R Documentation

Useful panel function with qqmath

Description

Useful panel function with qqmath. Draws a line passing through the points (usually) determined by the .25 and .75 quantiles of the sample and the theoretical distribution.

Usage

panel.qqmathline(x, y = x,
                 distribution = qnorm,
                 probs = c(0.25, 0.75),
                 qtype = 7,
                 groups = NULL, 
                 ...)

Arguments

x 	 The original sample, possibly reduced to a fewer number of quantiles, as determined by the f.value argument to qqmath

y 	 an alias for x for backwards compatibility

distribution 	 quantile function for reference theoretical distribution.

probs 	 numeric vector of length two, representing probabilities. Corresponding quantile pairs define the line drawn.

qtype 	 the type of quantile computation used in quantile

groups 	 optional grouping variable. If non-null, a line will be drawn for each group.

... 	 other arguments.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

prepanel.qqmathline, qqmath, quantile

[Package lattice version 0.17-25 Index]