panel.smoothScatter {lattice}	R Documentation

Lattice panel function analogous to smoothScatter

Description

This function allows the user to place smoothScatter plots in lattice graphics.

Usage

panel.smoothScatter(x, y = NULL,
                    nbin = 64, cuts = 255,
                    bandwidth,
                    colramp,
                    nrpoints = 100,
                    transformation = function(x) x^0.25,
                    pch = ".",
                    cex = 1, col="black",
                    range.x,
                    ...,
                    subscripts)

Arguments

x 	 Numeric vector containing x-values or n by 2 matrix containing x and y values.

y 	 Numeric vector containing y-values (optional). The length of x must be the same as that of y.

nbin 	 Numeric vector of length 1 (for both directions) or 2 (for x and y separately) containing the number of equally spaced grid points for the density estimation.

cuts 	 number of cuts defining the color gradient

bandwidth 	 Numeric vector: the smoothing bandwidth. If missing, these functions come up with a more or less useful guess. This parameter then gets passed on to the function bkde2D.

colramp 	 Function accepting an integer n as an argument and returning n colors.

nrpoints 	 Numeric vector of length 1 giving number of points to be superimposed on the density image. The first nrpoints points from those areas of lowest regional densities will be plotted. Adding points to the plot allows for the identification of outliers. If all points are to be plotted, choose nrpoints = Inf.

transformation 	 Function that maps the density scale to the color scale.

pch, cex 	 graphical parameters for the nrpoints “outlying” points shown in the display

range.x 	 see bkde2D for details.

col 	 points color parameter

... 	 Further arguments that are passed on to panel.levelplot.

subscripts 	 ignored, but necessary for handling of ... in certain situations. Likely to be removed in future.

Details

This replicates the display part of the smoothScatter function by replacing standard graphics calls by grid-compatible ones.

Value

The function is called for its side effects, namely the production of the appropriate plots on a graphics device.

Author(s)

Deepayan Sarkar deepayan.sarkar@r-project.org

Examples

ddf <- as.data.frame(matrix(rnorm(40000), ncol = 4) + 3 * rnorm(10000))
ddf[, c(2,4)] <- (-ddf[, c(2,4)])
xyplot(V1 ~ V2 + V3, ddf, outer = TRUE,
       panel = panel.smoothScatter, aspect = "iso")
splom(ddf, panel = panel.smoothScatter, nbin = 64)

[Package lattice version 0.17-25 Index]