panel.stripplot {lattice}	R Documentation

Default Panel Function for stripplot

Description

This is the default panel function for stripplot. Also see panel.superpose

Usage

panel.stripplot(x, y, jitter.data = FALSE,
                factor = 0.5, amount = NULL,
                horizontal = TRUE, groups = NULL,
                ...)

Arguments

x,y 	 coordinates of points to be plotted

jitter.data 	 whether points should be jittered to avoid overplotting. The actual jittering is performed inside panel.xyplot, using its jitter.x or jitter.y argument (depending on the value of horizontal).

factor, amount 	 amount of jittering, see jitter

horizontal 	 logical. If FALSE, the plot is ‘transposed’ in the sense that the behaviours of x and y are switched. x is now the ‘factor’. Interpretation of other arguments change accordingly. See documentation of bwplot for a fuller explanation.

groups 	 optional grouping variable

... 	 additional arguments, passed on to panel.xyplot

Details

Creates stripplot (one dimensional scatterplot) of x for each level of y (or vice versa, depending on the value of horizontal)

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

stripplot, jitter

[Package lattice version 0.17-25 Index]