panel.superpose {lattice}	R Documentation

Panel Function for Display Marked by groups

Description

These are panel functions for Trellis displays useful when a grouping variable is specified for use within panels. The x (and y where appropriate) variables are plotted with different graphical parameters for each distinct value of the grouping variable.

Usage

panel.superpose(x, y = NULL, subscripts, groups,
                panel.groups = "panel.xyplot",
                ...,
                col, col.line, col.symbol,
                pch, cex, fill, font,
                fontface, fontfamily,
                lty, lwd, alpha,
                type = "p",
                distribute.type = FALSE)
panel.superpose.2(..., distribute.type = TRUE)

Arguments

x,y 	 coordinates of the points to be displayed

panel.groups 	 the panel function to be used for each group of points. Defaults to panel.xyplot (behaviour in S).
To be able to distinguish between different levels of the originating group inside panel.groups, it will be supplied a special argument called group.number which will hold the numeric code corresponding to the current level of groups. No special care needs to be taken when writing a panel.groups function if this feature is not used.

subscripts 	 subscripts giving indices in original data frame

groups 	 a grouping variable. Different graphical parameters will be used to plot the subsets of observations given by each distinct value of groups. The default graphical parameters are obtained from superpose.symbol and superpose.line using trellis.par.get wherever appropriate

type 	 usually a character vector specifying what should be drawn for each group, passed on to the panel.groups function, which must know what to do with it. By default, this is panel.xyplot, whose help page describes the admissible values.
The functions panel.superpose and panel.superpose.2 differ only in the default value of distribute.type, which controls the way the type argument is interpreted. If distribute.type = FALSE, then the interpretation is the same as for panel.xyplot for each of the unique groups. In other words, if type is a vector, all the individual components are honoured concurrently. If distribute.type = TRUE, type is replicated to be as long as the number of unique values in groups, and one component used for the points corresponding to the each different group. Even in this case, it is possible to request multiple types per group, specifying type as a list, each component being the desired type vector for the corresponding group.
If distribute.type = FALSE, any occurrence of "g" in type causes a grid to be drawn, and all such occurrences are removed before type is passed on to panel.groups.

col, col.line, col.symbol, pch, cex, fill, font, fontface, fontfamily, lty, lwd, alpha 	 graphical parameters, replicated to be as long as the number of groups. These are eventually passed down to panel.groups, but as scalars rather than vectors. When panel.groups is called for the i-th level of groups, the corresponding element of each graphical parameter is passed to it.

... 	 Extra arguments. Passed down to panel.superpose from panel.superpose.2, and to panel.groups from panel.superpose.

distribute.type 	 logical controlling interpretation of the type argument.

Details

panel.superpose and panel.superpose.2 differ essentially in how type is interpreted by default. The default behaviour in panel.superpose is the opposite of that in S, which is the same as that of panel.superpose.2.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org (panel.superpose.2 originally contributed by Neil Klepeis)

See Also

Different functions when used as panel.groups gives different types of plots, for example panel.xyplot, panel.dotplot and panel.linejoin (This can be used to produce interaction plots).

See Lattice for an overview of the package.

[Package lattice version 0.17-25 Index]