F_3_prepanel.default {lattice}	R Documentation

Default Prepanel Functions

Description

These prepanel functions are used as fallback defaults in various high level plot functions in Lattice. These are rarely useful to normal users but may be helpful in developing new displays.

Usage



prepanel.default.bwplot(x, y, horizontal, nlevels, origin, stack, ...)
prepanel.default.histogram(x, breaks, equal.widths, type, nint, ...)
prepanel.default.qq(x, y, ...)
prepanel.default.xyplot(x, y, type, subscripts, groups, ...)
prepanel.default.cloud(perspective, distance,
                       xlim, ylim, zlim,
                       screen = list(z = 40, x = -60),
                       R.mat = diag(4),
                       aspect = c(1, 1), panel.aspect = 1,
                       ..., zoom = 0.8)
prepanel.default.levelplot(x, y, subscripts, ...)
prepanel.default.qqmath(x, f.value, distribution, qtype,
                        groups, subscripts, ..., tails.n = 0)
prepanel.default.densityplot(x, darg, groups, weights, subscripts, ...)
prepanel.default.parallel(x, y, z, ..., horizontal.axis)
prepanel.default.splom(z, ...)


Arguments

x, y
x and y values, numeric or factor

horizontal
logical, applicable when one of the variables is to be treated as categorical (factor or shingle).

horizontal.axis
logical indicating whether the parallel axes should be laid out horizontally (TRUE) or vertically (FALSE).

nlevels
number of levels of such a categorical variable.

origin, stack
for barcharts or the type="h" plot type

breaks, equal.widths, type, nint
details of histogram calculations. type has a different meaning in prepanel.default.xyplot (see panel.xyplot)

groups, subscripts
See xyplot. Whenever appropriate, calculations are done separately for each group and then combined.

weights
numeric vector of weights for the density calculations. If this is specified, it is subsetted by subscripts to match it to x.

perspective, distance, xlim, ylim, zlim, screen, R.mat, aspect, panel.aspect, zoom
see panel.cloud

f.value, distribution, tails.n
see panel.qqmath

darg
list of arguments passed to density

z
see panel.parallel and panel.pairs

qtype
type of quantile

...
other arguments, usually ignored

Value

A list with components xlim, ylim, dx and dy, and possibly xat and yat, the first two being used to calculate panel axes limits, the last two for banking computations. The form of these components are described in the help page for xyplot.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

xyplot, banking, Lattice. See documentation of corresponding panel functions for more details about the arguments.

[Package lattice version 0.20-35 Index]