prepanel.functions {lattice}	R Documentation

Useful Prepanel Function for Lattice

Description

These are predefined prepanel functions available in Lattice.

Usage

prepanel.lmline(x, y, ...)
prepanel.loess(x, y, span, degree, family, evaluation, ...)
prepanel.qqmathline(x, y = x, distribution = qnorm,
                    probs = c(0.25, 0.75), qtype = 7,
                    groups, subscripts,
                    ...) 

Arguments

x, y 	 x and y values, numeric or factor

distribution 	 quantile function for theoretical distribution. This is automatically passed in when this is used as a prepanel function in qqmath.

qtype 	 type of quantile

probs 	 numeric vector of length two, representing probabilities. If used with aspect="xy", the aspect ratio will be chosen to make the line passing through the corresponding quantile pairs as close to 45 degrees as possible.

span, degree, family, evaluation 	 arguments controlling the underlying loess smooth

groups, subscripts 	 See xyplot. Whenever appropriate, calculations are done separately for each group and then combined.

... 	 other arguments

Value

usually a list with components xlim, ylim, dx and dy, the first two being used to calculate panel axes limits, the last two for banking computations. The form of these components are described under xyplot. There are also several prepanel functions that serve as the default for high level functions, see prepanel.default.xyplot

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

trellis.par.get, xyplot, banking, Lattice. See loess.smooth for further options to prepanel.loess

[Package lattice version 0.17-25 Index]