qq {lattice}	R Documentation

Quantile-Quantile Plots of Two Samples

Description

Quantile-Quantile plots for comparing two Distributions

Usage

qq(x, data, ...)

## S3 method for class 'formula':
qq(x, data, aspect = "fill", 
    panel = lattice.getOption("panel.qq"),
    prepanel, scales, strip, 
    groups, xlab, xlim, ylab, ylim, f.value = NULL, 
    drop.unused.levels = lattice.getOption("drop.unused.levels"),
    ...,
    lattice.options = NULL,
    qtype = 7,
    default.scales = list(),
    subscripts,
    subset)

Arguments

x 	 The object on which method dispatch is carried out.
For the formula method, a formula of the form y ~ x | g1 * g2 * ..., where x must be a numeric, and y can be a factor, shingle, character or numeric vector, with the restriction that there must be exactly two levels of y, which divide the values of x into two groups. Quantiles for these groups will be plotted along the two axes.

data 	 For the formula methods, an optional data frame in which variables in the formula (as well as groups and subset, if any) are to be evaluated.

f.value 	 optional numeric vector of probabilities, quantiles corresponding to which should be plotted. Can also be a function of a single integer (representing sample size) that returns such a numeric vector. The typical value for this argument is the function ppoints, which is also the S-PLUS default. If specified, the probabilities generated by this function is used for the plotted quantiles, using the quantile function.
f.value defaults to NULL, which is equivalent to using function(n) ppoints(n, a = 1). This has the effect of including the minimum and maximum data values in the computed quantiles. This is similar to what happens for qqplot but different from the default qq behaviour in S-PLUS.
For large data, this argument can be useful in plotting a smaller set of quantiles, which is usually enough to capture the pattern.

panel 	 The function that uses the packet (subset of display variables) corresponding to a panel to create a display. Default panel functions are documented separately, and often have arguments that can be used to customize its display in various ways. Such arguments can usually be directly supplied to the high level function.

qtype 	 type argument for the quantile

aspect, prepanel, scales, strip, groups, xlab, xlim, ylab, ylim, drop.unused.levels, lattice.options, default.scales, subscripts, subset 	 See xyplot

... 	 Further arguments. See corresponding entry in xyplot for non-trivial details.

Details

qq produces a Q-Q plot of two samples. The default behaviour of qq is different from the corresponding S-PLUS function. See the entry for f.value for specifics.

This and all other high level Trellis functions have several arguments in common. These are extensively documented only in the help page for xyplot, which should be consulted to learn more detailed usage.

Value

An object of class "trellis". The update method can be used to update components of the object and the print method (usually called by default) will plot it on an appropriate plotting device.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

xyplot, panel.qq, qqmath, Lattice

Examples

qq(voice.part ~ height, aspect = 1, data = singer,
   subset = (voice.part == "Bass 2" | voice.part == "Tenor 1"))

[Package lattice version 0.17-25 Index]