B_10_rfs {lattice}	R Documentation

Residual and Fit Spread Plots

Description

Plots fitted values and residuals (via qqmath) on a common scale for any object that has methods for fitted values and residuals.

Usage


rfs(model, layout=c(2, 1), xlab="f-value", ylab=NULL,
    distribution = qunif,
    panel, prepanel, strip, ...)

Arguments

model
a fitted model object with methods fitted.values and residuals. Can be the value returned by oneway

layout
default layout is c(2,1)

xlab
defaults to "f.value"

distribution
the distribution function to be used for qqmath

ylab, panel, prepanel, strip
See xyplot

...
other arguments, passed on to qqmath.

Value

An object of class "trellis". The update method can be used to update components of the object and the print method (usually called by default) will plot it on an appropriate plotting device.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

oneway, qqmath, xyplot, Lattice

Examples


rfs(oneway(height ~ voice.part, data = singer, spread = 1), aspect = 1)

[Package lattice version 0.20-35 Index]