tmd {lattice}	R Documentation

Tukey Mean-Difference Plot

Description

Create Tukey Mean-Difference Plots from a trellis object returned by xyplot and qq

Usage

tmd(object,
    xlab = "mean",
    ylab = "difference",
    panel = "panel.tmd",
    prepanel = "prepanel.default.tmd",
    ...)

Arguments

object 	 An object of class ``trellis'' returned by xyplot or qq

xlab 	 x label

ylab 	 y label

panel 	 panel function to be used

prepanel 	 prepanel function used to recalculate limits using the modified x and y values

... 	 other arguments

Details

The Tukey Mean-difference plot is produced by modifying the (x,y) values of each panel as follows: the new coordinates are given by x=(x+y)/2 and y=y-x.

tmd uses the update method for ``trellis'' objects, which processes all extra arguments supplied to tmd.

Value

An object of class ``trellis''. The `update' method can be used to update components of the object and the `print' method (usually called by default) will plot it on an appropriate plotting device.

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

panel.tmd, xyplot, Lattice

Examples

tmd(qqmath(~height|voice.part, data = singer))

[Package lattice version 0.11-6 Index]