D_trellis.object {lattice}	R Documentation

A Trellis Plot Object

Description

This class of objects is returned by high level lattice functions, and is usually plotted by default by its print method.

Details

A trellis object, as returned by high level lattice functions like xyplot, is a list with the "class" attribute set to "trellis". Many of the components of this list are simply the arguments to the high level function that produced the object. Among them are: as.table, layout, page, panel, prepanel, main, sub, par.strip.text, strip, skip, xlab ylab, par.settings, lattice.options and plot.args. Some other typical components are:

formula

the Trellis formula used in the call

index.cond

list with index for each of the conditioning variables

perm.cond

permutation of the order of the conditioning variables

aspect.fill

logical, whether aspect is "fill"

aspect.ratio

numeric, aspect ratio to be used if aspect.fill is FALSE

call

call that generated the object.

condlevels

list with levels of the conditioning variables

legend

list describing the legend(s) to be drawn

panel.args

a list as long as the number of panels, each element being a list itself, containing the arguments in named form to be passed to the panel function in that panel.

panel.args.common

a list containing the arguments common to all the panel functions in name=value form

x.scales

list describing x-scale, can consist of several other lists, paralleling panel.args, if x-relation is not "same"

y.scales

list describing y-scale, similar to x.scales

x.between

numeric vector of interpanel x-space

y.between

numeric vector of interpanel y-space

x.limits

numeric vector of length 2 or list, giving x-axis limits

y.limits

similar to x.limits

packet.sizes

array recording the number of observations in each packet

Author(s)

Deepayan Sarkar Deepayan.Sarkar@R-project.org

See Also

Lattice, xyplot, print.trellis

[Package lattice version 0.20-35 Index]